objprint: the current object matrix is a MtxPtr, and the fuzz distance reads its translation row

+++ src/dlls/objects/198_AnimatedObj/AnimatedObj.c
@@ -89,7 +89,7 @@ void animatedobj_render(GameObject* obj, int fwdArg2, int fwdArg3, int fwdArg4, 
         PSMTXConcat(flipZ, cameraFlippedY, cameraFlippedYZ);
         PSMTXConcat(cameraTranslation, cameraFlippedYZ, mirroredCamera);
         PSMTXConcat(mirroredCamera, translatedWorld, renderMatrix);
-        objSetCurrentMatrix((u32)renderMatrix);
+        objSetCurrentMatrix(renderMatrix);
         objRenderModel(obj);
     } else {
         objRenderModelAndHitVolumes(obj, fwdArg2, fwdArg3, fwdArg4, fwdArg5, 1.0f);
